1. fix model - Sushi for init data 2. add iOS 13 window root controller use AppDelegate and UINavigation 3. add new viewController files for my Menu - ContainerController, MenuController
