+++ Sushiwok/AppDelegate.swift
@@ -15,11 +15,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
-        
-        window = UIWindow()
-        window?.makeKeyAndVisible()
-        window?.rootViewController = ContainerViewController()
-        
         return true
     }
